onlineShopWebsiteProjectDjango-with-db.sqlite3/requirements.txt
asgiref==3.9.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
distlib==0.3.9
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==4.2.23
django-allauth==65.9.0
django-cache-url==3.4.5
django-crispy-forms==2.4
django-jalali-date==2.0.0
django-rosetta==0.10.2
environs==14.2.0
filelock==3.18.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
marshmallow==4.0.0
packaging==25.0
pillow==11.2.1
pipenv==2025.0.2
platformdirs==4.3.8
polib==1.2.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
